created new file distgnn_kernel.cc
