+++ python/dgl/distgnn/drpa_sym.py
@@ -1,10 +1,10 @@
 """
 Copyright (c) 2021 Intel Corporation
  \file distgnn/partition/main_Libra.py
- \brief Libra - Vertex-cut based graph partitioner for distirbuted training
+ \brief Libra - Vertex-cut based graph partitioner for distributed training
  \author Vasimuddin Md <vasimuddin.md@intel.com>,
          Sanchit Misra <sanchit.misra@intel.com>,
-         Ramanarayan Mohanty <ramanarayan.mohanty@intel.com>,         
+         Ramanarayan Mohanty <ramanarayan.mohanty@intel.com>,
          Sasikanth Avancha <sasikanth.avancha@intel.com>
 """
 
@@ -12,8 +12,7 @@ import torch
 import time
 from torch import nn
 from torch.nn import functional as F
-import os, psutil
-import gc
+import os
 import sys
 
 from .. import function as fn
@@ -23,7 +22,7 @@ from .. import DGLHeteroGraph
 from dgl.base import DGLError
 from ..sparse import scatter_reduce_lr, fdrpa_gather_emb_lr
 from ..sparse import scatter_reduce_rl, fdrpa_gather_emb_rl
-from ..sparse import fdrpa_comm_buckets, deg_div, deg_div_back, fdrpa_init_buckets
+from ..sparse import fdrpa_comm_buckets
 
 display = False
 class GQueue():
@@ -53,7 +52,7 @@ class GQueue():
     def purge(self):
         self.queue = []
 
-        
+
 ## communication storage
 gfqueue_feats_lr  = GQueue()
 gfqueue_nodes_lr = GQueue()
@@ -93,7 +92,7 @@ def drpa(gobj, rank, num_parts, node_map, nrounds, dist, nlayers):
     return d
 
 
-class DRPAMaster(DGLHeteroGraph):            
+class DRPAMaster(DGLHeteroGraph):
     def drpa_init(self, rank, num_parts, node_map, nrounds, dist, nlayers):
         self.rank = rank
         self.num_parts = num_parts
@@ -104,19 +103,15 @@ class DRPAMaster(DGLHeteroGraph):
 
         self.epochs_ar = [0 for i in range(self.nlayers)]
         self.epochi = 0
-        #self.gather_q41 = GQueue()  #TODOs
-        #self.output_sr_ar = []    #TODOs
-
+        
         if self.nrounds == -1: return     ## 0C algoriithm, no drpa applicable
 
         adj = self.dstdata['adj']
         lf = self.dstdata['lf']
         width = adj.shape[1]
 
-        ## Creates buckets based on ndrounds        
+        ## Creates buckets based on ndrounds
         self.drpa_create_buckets()
-        # self.drpa_init_buckets(adj, lf, width)  TODOs
-
         
     def drpa_finalize(self):
         if self.rank == 0:
@@ -126,7 +121,7 @@ class DRPAMaster(DGLHeteroGraph):
             req.wait()
             req = gfqueue_nodes_lr.pop()
             req.wait()
-        while gfqueue_feats_rl.empty() == False:                
+        while gfqueue_feats_rl.empty() == False:
             req = gfqueue_feats_rl.pop()
             req.wait()
 
@@ -158,7 +153,6 @@ class DRPAMaster(DGLHeteroGraph):
                    reduce_func,
                    apply_node_func=None,
                    etype=None):
-        #assert self.rank != -1, "drpa not initialized !!!"
         if self.rank == -1:
             raise DGLError("Error: drpa not initialized!")
 
@@ -179,18 +173,18 @@ class DRPAMaster(DGLHeteroGraph):
             if mean == 1:
                 feat_dst = self.dstdata['h']
                 self.r_in_degs = DGLHeteroGraph.in_degrees(self).to(feat_dst)
-                self.dstdata['neigh'] = self.dstdata['neigh'] / self.r_in_degs.unsqueeze(-1)            
+                self.dstdata['neigh'] = self.dstdata['neigh'] / self.r_in_degs.unsqueeze(-1)
             return
         
         neigh = self.dstdata['neigh']
         adj = self.dstdata['adj']
         inner_node = self.dstdata['inner_node']
         lftensor = self.dstdata['lf']
-        feat_dst = self.dstdata['h']        
+        feat_dst = self.dstdata['h']
         epoch = self.epochs_ar[self.epochi]
 
         ## Remote Aggregate
-        tic = time.time()        
+        tic = time.time()
         self.r_in_degs = DGLHeteroGraph.in_degrees(self).to(feat_dst)
         self.dstdata['neigh'] = call_drpa_core(neigh, adj, inner_node,
                                                lftensor, self.selected_nodes,
@@ -205,37 +199,17 @@ class DRPAMaster(DGLHeteroGraph):
 
         
         toc = time.time()
-        if self.rank == 0 and display:        
+        if self.rank == 0 and display:
             print("Time for remote aggregate: {:0.4f}".format(toc - tic))
 
         if mean == 1:
             self.dstdata['neigh'] = self.dstdata['neigh'] / self.r_in_degs.unsqueeze(-1)
+
             
-
-    #def drpa_init_buckets(self, adj, lf, width):
-    #
-    #    for l in range(self.nrounds):
-    #        buckets = torch.tensor([0 for i in range(self.num_parts)], dtype=torch.int32)
-    #        sn = torch.tensor(self.selected_nodes[l], dtype=torch.int32)
-    #        nm = torch.tensor(self.node_map, dtype=torch.int32)
-    #        fdrpa_init_buckets(adj, sn, nm, buckets,
-    #                           lf, width, self.num_parts, self.rank)
-    #        input_sr = []
-    #        for i in range(0, self.num_parts):
-    #            input_sr.append(torch.tensor([buckets[i]], dtype=torch.int64))
-    #            
-    #        output_sr = [torch.zeros(1, dtype=torch.int64) for i in range(0, self.num_parts)]
-    #        sync_req = self.dist.all_to_all(output_sr, input_sr, async_op=True)  
-    #        sync_req.wait() ## recv the #nodes communicated
-    #        self.output_sr_ar.append(output_sr)  ## output
-
-
     def drpa_create_buckets(self):
         inner_nodex = self.ndata['inner_node'].tolist() ##.count(1)
         n = len(inner_nodex)
         idx = inner_nodex.count(1)
-        
-        #nrounds_ = self.nrounds
         self.selected_nodes = [ [] for i in range(self.nrounds)]  ## #native nodes
 
         # randomly divide the nodes in 5 rounds for comms
@@ -254,13 +228,10 @@ class DRPAMaster(DGLHeteroGraph):
                     
             counter += 1
             
-        if (counter != len(inner_nodex)):
-            print("counter: ", counter, " ", len(inner_nodex))            
-        #assert counter == len(inner_nodex), "assertion"
         if counter != len(inner_nodex):
-            raise DGLError("Error: Issue in selected nodes.")
+            raise DGLError("Error: Issue in selected nodes, counter not matching.")
 
-                
+
     def in_degrees(self):
         try:
             return self.r_in_degs
@@ -271,7 +242,7 @@ class DRPAMaster(DGLHeteroGraph):
         return DGLHeteroGraph.in_degrees(self)
 
 
-def message(rank, msg, val=-1, val_=-1):    
+def message(rank, msg, val=-1, val_=-1):
     if rank == 0 and display:
         if val == -1:
             print(msg, flush=True)
@@ -280,11 +251,11 @@ def message(rank, msg, val=-1, val_=-1):
         else:
             print(msg.format(val, val_), flush=True)
 
-            
+
 class DRPACore(torch.autograd.Function):
     @staticmethod
     def forward(ctx, feat, adj, inner_node, lftensor, selected_nodes,
-                node_map, num_parts, rank, epoch, dist, in_degs, nrounds):        
+                node_map, num_parts, rank, epoch, dist, in_degs, nrounds):
         """
         feat : graph (partition node features)
         adj : list of remote clone nodes
@@ -304,19 +275,19 @@ class DRPACore(torch.autograd.Function):
         nrounds_update = nrounds
         feat_size = feat.shape[1]
         int_threshold = pow(2, 31)/4 - 1              ##bytes
-        base_chunk_size = int(int_threshold / num_parts) ##bytes            
+        base_chunk_size = int(int_threshold / num_parts) ##bytes
         base_chunk_size_fs = floor(base_chunk_size / (feat_size + 1) )
         roundn =  epoch % nrounds
         
         node_map_t = torch.tensor(node_map, dtype=torch.int32)
         selected_nodes_t = []
         for sn in selected_nodes:
-            selected_nodes_t.append(torch.tensor(sn, dtype=torch.int32))            
+            selected_nodes_t.append(torch.tensor(sn, dtype=torch.int32))
             
         buckets = torch.tensor([0 for i in range(num_parts)], dtype=torch.int32)
         width = adj.shape[1]   ## feature vector length
 
-        tic = time.time()
+        #tic = time.time()
         # 1. Get bucket sizes to be sent to all the ranks
         #    And get those remote nodes locations
         ver2part = torch.empty(len(selected_nodes[roundn]), dtype=torch.int32)
@@ -324,7 +295,7 @@ class DRPACore(torch.autograd.Function):
         fdrpa_comm_buckets(adj, selected_nodes_t[roundn], ver2part, ver2part_index,
                            node_map_t, buckets, lftensor, width, num_parts, rank)
         
-        message(rank, "Time for bucketing: {:0.4f}", (time.time() - tic))        
+        #message(rank, "Time for bucketing: {:0.4f}", (time.time() - tic))
             
         ## comms to gather the bucket sizes for all-to-all feats comms
         input_sr = []
@@ -332,7 +303,7 @@ class DRPACore(torch.autograd.Function):
             input_sr.append(torch.tensor([buckets[i]], dtype=torch.int64))
 
         output_sr = [torch.zeros(1, dtype=torch.int64) for i in range(0, num_parts)]
-        sync_req = dist.all_to_all(output_sr, input_sr, async_op=True)  
+        sync_req = dist.all_to_all(output_sr, input_sr, async_op=True)
         sync_req.wait() ## recv the #nodes communicated
             
         send_feat_len = 0
@@ -340,7 +311,7 @@ class DRPACore(torch.autograd.Function):
         for i in range(num_parts):
             in_size.append(int(buckets[i]) * (feat_size + 1))
             send_feat_len += in_size[i]
-                     
+            
         ## 2. Split the data if the communication volume is beyond MPI limit
         ##############################################################################
         tic = time.time()            
@@ -353,7 +324,7 @@ class DRPACore(torch.autograd.Function):
             if int(i) >= base_chunk_size_fs: flg = 1
         
         nsplit_comm = 1
-        if cum >= int_threshold or send_feat_len >= int_threshold or flg:        
+        if cum >= int_threshold or send_feat_len >= int_threshold or flg:
             for i in range(num_parts):
                 val = ceil((int(input_sr[i]) ) / base_chunk_size_fs)
                 if val > nsplit_comm: nsplit_comm = val
@@ -364,8 +335,8 @@ class DRPACore(torch.autograd.Function):
                                                        op=torch.distributed.ReduceOp.MAX,
                                                        async_op=True)
         
-        lim = 1            
-        soffset_base = [0 for i in range(num_parts)]  ## min chunk size        
+        lim = 1
+        soffset_base = [0 for i in range(num_parts)]  ## min chunk size
         soffset_cur = [0 for i in range(num_parts)]       ## send list of ints
         roffset_cur = [0 for i in range(num_parts)]       ## recv list of intrs
         
@@ -405,7 +376,7 @@ class DRPACore(torch.autograd.Function):
 
             offset = 0
             ## 2.1 gather feats to send by leaf nodes
-            for i in range(num_parts): 
+            for i in range(num_parts):
                 fdrpa_gather_emb_lr(feat, feat.shape[0], adj, sten_, offset,
                                     send_node_list[i], sten_nodes,
                                     selected_nodes_t[roundn],
@@ -421,11 +392,11 @@ class DRPACore(torch.autograd.Function):
             message(rank, "Sending {}, recving {} data I", tsend, trecv)
 
             ## Communicate the features and node IDs
-            req_feats = dist.all_to_all_single(dten_, sten_, out_size, in_size, async_op=True)    	
+            req_feats = dist.all_to_all_single(dten_, sten_, out_size, in_size, async_op=True)
             gfqueue_feats_lr.push(req_feats)
             req_nodes = dist.all_to_all_single(dten_nodes, sten_nodes,
                                          out_size_nodes, in_size_nodes,
-                                         async_op=True)    	
+                                         async_op=True)
             gfqueue_nodes_lr.push(req_nodes)
                 
             soffset_cur_copy = soffset_cur.copy()
@@ -442,17 +413,15 @@ class DRPACore(torch.autograd.Function):
                 req_nsplit_comm.wait()
                 lim = int(nsplit_comm_t)
             j += 1
-        ##############################################################################        
+        ##############################################################################
         buffcomm_iter_lr.push(lim)
 
-        message(rank, "Max iters in MPI split comm: {}", (lim))                    
-        #message(rank, "Time for Gather I: {:0.4f}", (time.time() - tic))
-        prof.append('Gather I: {:0.4f}'.format(time.time() - tic))            
+        #message(rank, "Max iters in MPI split comm: {}", (lim))
+        prof.append('Gather I: {:0.4f}'.format(time.time() - tic))
         
         ## 3. Recv the remote partial aggreates by the root and update the aggregates
-        recv_list_nodes = []        
+        recv_list_nodes = []
         if epoch >= nrounds_update or nrounds == 1:
-            #assert gfqueue_feats_lr.empty() == False, "Error: Forward empty queue !!!"
             if gfqueue_feats_lr.empty() == True:
                 raise DGLError("Error: unexpected event, forward empty queue.")
 
@@ -460,12 +429,11 @@ class DRPACore(torch.autograd.Function):
             lim = buffcomm_iter_lr.pop()
             out_size_nodes_ar = []
             
-            for i in range(lim):                
+            for i in range(lim):
                 if rank == 0 and display: tic = time.time()
                 req = gfqueue_feats_lr.pop();  req.wait()
                 req = gfqueue_nodes_lr.pop(); req.wait()
             
-                #message(rank, "Time for async comms I: {:4f}", (time.time() - tic))
                 prof.append('Async comm I: {:0.4f}'.format(time.time() - tic))
                     
                 otf = buffcomm_feats_lr.pop()
@@ -479,7 +447,7 @@ class DRPACore(torch.autograd.Function):
                     ilen += out_size_nodes[l]
                     recv_list_nodes_ar.append(torch.empty(out_size_nodes[l], dtype=torch.int32))
             
-                pos = torch.tensor([0], dtype=torch.int64)        
+                pos = torch.tensor([0], dtype=torch.int64)
                 offsetf = 0; offsetn = 0
                 ## 3.1 Scatter: update the local aggregates using remote recv aggregates
                 for l in range(num_parts):
@@ -491,13 +459,10 @@ class DRPACore(torch.autograd.Function):
                     offsetf += out_size[l]
                     offsetn += out_size_nodes[l]
                     
-                #assert ilen == pos[0], "Issue in scatter reduce!"
                 if ilen != pos[0]: raise DGLError("Error: Issue in scatter reduce.")
                 recv_list_nodes.append(recv_list_nodes_ar)
 
-
-            #message(rank, "Time for scatter I: {:0.4f} in epoch: {}", (time.time() - ticg), epoch)
-            prof.append('Scatter I: {:0.4f}'.format(time.time() - tic))
+            prof.append('Scatter I: {:0.4f}'.format(time.time() - ticg))
 
             tic = time.time()
             for j in range(lim):               ### gather-scatter round II
@@ -509,9 +474,9 @@ class DRPACore(torch.autograd.Function):
                     tsend += out_size_nodes[i]
                     trecv += stn_fp2[i]
                     
-                recv_list_nodes_ = recv_list_nodes[j]                    
+                recv_list_nodes_ = recv_list_nodes[j]
                 sten_ = torch.empty(tsend * (feat_size + 1), dtype=feat.dtype)
-                dten_ = torch.empty(trecv * (feat_size + 1), dtype=feat.dtype)            
+                dten_ = torch.empty(trecv * (feat_size + 1), dtype=feat.dtype)
                 
                 out_size = [0 for i in range(num_parts)]
                 in_size = [0 for i in range(num_parts)]
@@ -526,28 +491,26 @@ class DRPACore(torch.autograd.Function):
                     out_size[i]       = stn_fp2[i] * (feat_size + 1)
                     in_size[i]        = out_size_nodes[i] * (feat_size + 1)
                     offset           += in_size[i]
-                                    
-                req = dist.all_to_all_single(dten_, sten_, out_size, in_size, async_op=True)    	
-                gfqueue_feats_rl.push(req)                
+                    
+                req = dist.all_to_all_single(dten_, sten_, out_size, in_size, async_op=True)
+                gfqueue_feats_rl.push(req)
                 ## push dten
                 buffcomm_feats_rl.push(dten_)
                 buffcomm_feats_size_rl.push(out_size)
                     
             buffcomm_iter_rl.push(lim)
 
-            #message(rank, "Time for gather 2: {:0.4f}",(time.time() - tic))
             prof.append('Gather II: {:0.4f}'.format(time.time() - tic))
 
             ## 4. Recv remote partial aggreagted from the root and update local aggregates
             if epoch >= 2*nrounds_update or nrounds == 1:
                 ticg = time.time()
                     
-                lim = buffcomm_iter_rl.pop()                    
+                lim = buffcomm_iter_rl.pop()
                 for i in range(lim):
                     tic = time.time()
                     
                     req = gfqueue_feats_rl.pop(); req.wait()
-                    #message(rank, "Time for async comms II: {:4f}", (time.time() - tic))
                     prof.append('Async comms II: {:0.4f}'.format(time.time() - tic))
                     
                     otf = buffcomm_feats_rl.pop()
@@ -563,13 +526,12 @@ class DRPACore(torch.autograd.Function):
             
                         offset += out_size[l]
 
-                #message(rank, "Time for scatter 2: {:0.4f}, roundn: {}", time.time() - ticg, roundn)
                 prof.append('Scatter II: {:0.4f}'.format(time.time() - ticg))
 
         if rank == 0:  ## Display runtime profile for major components
             print(prof, flush=True)
-            print()
-                
+            #print()
+            
         return feat
 
 
@@ -579,7 +541,6 @@ class DRPACore(torch.autograd.Function):
             None, None, None, None, None, None, None
 
 
-    
 def call_drpa_core(neigh, adj,inner_node, lftensor, selected_nodes, node_map,
                    num_parts, rank, epoch, dist, degs, nrounds):
     return DRPACore.apply(neigh, adj, inner_node, lftensor, selected_nodes,
