+++ python/dgl/sparse.py
@@ -603,23 +603,10 @@ def _csrmask(A, A_weights, B):
 
 ##################################################
 ## DistGNN functions
-def deg_div(neigh, h, feat_size, degs, lim):
-    _CAPI_DGLKernelDegDiv(to_dgl_nd_for_write(neigh),
-                          to_dgl_nd(h),
-                          feat_size,
-                          to_dgl_nd(degs),
-                          lim)
-    
-def deg_div_back(neigh, feat_size, degs, lim):
-    _CAPI_DGLKernelDegDivBack(to_dgl_nd_for_write(neigh),
-                              feat_size,
-                              to_dgl_nd(degs),
-                              lim)
-
 def fdrpa_gather_emb_lr(feat, feat_shape, adj, send_feat_list, offset,
                         send_node_list, send_to_node_list, selected_nodes,
                         in_degs, ver2part, ver2part_index, width, feat_size,
-                        cur_part, soffset_base, soffset_cur, node_map, num_parts):    
+                        cur_part, soffset_base, soffset_cur, node_map, num_parts):
     sfl_a            = to_dgl_nd_for_write(send_feat_list)
     feat_a           = to_dgl_nd(feat)
     adj_a            = to_dgl_nd(adj)
@@ -627,7 +614,6 @@ def fdrpa_gather_emb_lr(feat, feat_shape, adj, send_feat_list, offset,
     stnl_a           = to_dgl_nd_for_write(send_to_node_list)
     selected_nodes_a = to_dgl_nd(selected_nodes)
     node_map_a       = to_dgl_nd(node_map)
-    
     _CAPI_DGLKernelFdrpaGatherEmbLR(feat_a,
                                     feat_shape,
                                     adj_a,
@@ -655,7 +641,7 @@ def scatter_reduce_lr(otf, offsetf, otn, offsetn, neigh, degs, node_map,
     _CAPI_DGLKernelScatterReduceLR(to_dgl_nd(otf),
                                    offsetf,
                                    to_dgl_nd(otn),
-                                   offsetn,                                   
+                                   offsetn,
                                    to_dgl_nd_for_write(neigh),
                                    to_dgl_nd_for_write(degs),
                                    node_map_a,
@@ -667,7 +653,7 @@ def scatter_reduce_lr(otf, offsetf, otn, offsetn, neigh, degs, node_map,
                                    count,
                                    cur_part)
 
-    
+
 def fdrpa_gather_emb_rl(feat, feat_shape, send_feat_list, offset, recv_list_nodes,
                         lim, in_degs, feat_size, node_map, num_parts):
     node_map_a = to_dgl_nd(node_map)
@@ -678,11 +664,11 @@ def fdrpa_gather_emb_rl(feat, feat_shape, send_feat_list, offset, recv_list_node
                                     to_dgl_nd(recv_list_nodes),
                                     lim,
                                     to_dgl_nd(in_degs),
-                                    feat_size,                                    
+                                    feat_size,
                                     node_map_a,
                                     num_parts)
 
-    
+
 def scatter_reduce_rl(otf, offset, stn, lim, in_degs, neigh, node_map, dim, feat_size,
                       num_parts):
     node_map_a = to_dgl_nd(node_map)
@@ -696,7 +682,6 @@ def scatter_reduce_rl(otf, offset, stn, lim, in_degs, neigh, node_map, dim, feat
                                    dim,
                                    feat_size,
                                    num_parts)
-    
 
 def fdrpa_comm_buckets(adj, selected_nodes, ver2part, ver2part_index, node_map,
                        buckets, lf, width, num_parts, cur_part):
@@ -712,89 +697,6 @@ def fdrpa_comm_buckets(adj, selected_nodes, ver2part, ver2part_index, node_map,
                                     width,
                                     num_parts,
                                     cur_part)
-
-def fdrpa_init_buckets(adj, selected_nodes, node_map, buckets, lf, width, num_parts, cur_part):
-    selected_nodes_a =  to_dgl_nd(selected_nodes)
-    node_map_a = to_dgl_nd(node_map)
-    _CAPI_DGLKernelFdrpaInitBuckets(to_dgl_nd(adj),
-                                    selected_nodes_a,
-                                    node_map_a,
-                                    to_dgl_nd_for_write(buckets),
-                                    to_dgl_nd(lf),
-                                    width,
-                                    num_parts,
-                                    cur_part)
-
-    
-###################################################################################################
-## Libra Graph Partition
-def libra_vertex_cut(nc, node_degree, edgenum_unassigned,
-                     community_weights, u, v, w, out, N, N_e, dataset):    
-    _CAPI_DGLLibraVertexCut(nc,
-                            to_dgl_nd_for_write(node_degree),
-                            to_dgl_nd_for_write(edgenum_unassigned),
-                            to_dgl_nd_for_write(community_weights),
-                            to_dgl_nd(u),
-                            to_dgl_nd(v),
-                            to_dgl_nd(w),
-                            to_dgl_nd_for_write(out),
-                            N,
-                            N_e,
-                            dataset)
-
-
-def libra2dgl_build_dict(a, b, indices, ldt_key, gdt_key, gdt_value, node_map,
-                         offset, nc, c, fsize, hash_nodes, dataset):
-    _CAPI_DGLLibra2dglBuildDict(to_dgl_nd_for_write(a),
-                                to_dgl_nd_for_write(b),
-                                to_dgl_nd_for_write(indices),
-                                to_dgl_nd_for_write(ldt_key),
-                                to_dgl_nd_for_write(gdt_key),
-                                to_dgl_nd_for_write(gdt_value),
-                                to_dgl_nd_for_write(node_map),
-                                to_dgl_nd_for_write(offset),
-                                nc,
-                                c,
-                                fsize,
-                                to_dgl_nd_for_write(hash_nodes),
-                                dataset)
-    
-def libra2dgl_build_adjlist(feat, gfeat, adj, inner_node, ldt, gdt_key,
-                            gdt_value, node_map, lf, lftensor, num_nodes,
-                            nc, c, feat_size, labels, trainm, testm, valm,
-                            glabels, gtrainm, gtestm, gvalm, feat_shape):    
-    _CAPI_DGLLibra2dglBuildAdjlist(to_dgl_nd(feat),
-                                   to_dgl_nd_for_write(gfeat),
-                                   to_dgl_nd_for_write(adj),
-                                   to_dgl_nd_for_write(inner_node),
-                                   to_dgl_nd(ldt),
-                                   to_dgl_nd(gdt_key),
-                                   to_dgl_nd(gdt_value),
-                                   to_dgl_nd(node_map),
-                                   to_dgl_nd_for_write(lf),
-                                   to_dgl_nd(lftensor),                                 
-                                   num_nodes,
-                                   nc,
-                                   c,
-                                   feat_size,
-                                   to_dgl_nd(labels),
-                                   to_dgl_nd(trainm),
-                                   to_dgl_nd(testm),
-                                   to_dgl_nd(valm),
-                                   to_dgl_nd_for_write(glabels),
-                                   to_dgl_nd_for_write(gtrainm),
-                                   to_dgl_nd_for_write(gtestm),
-                                   to_dgl_nd_for_write(gvalm),
-                                   feat_shape)
-    
-
-    
-def libra2dgl_set_lf(gdt_key, gdt_value, lftensor, nc, Nn):
-    _CAPI_DGLLibra2dglSetLF(to_dgl_nd(gdt_key),
-                            to_dgl_nd(gdt_value),
-                            to_dgl_nd_for_write(lftensor),
-                            nc,
-                            Nn)
 
 
 _init_api("dgl.sparse")
